Tried to start a tool to overlay images for maps
Added a beautiful 700x700 Valhalla map, as well as a tool (currently
running in driver) to chop it up into nice 100x100 squares for Charlie's
graphics... thing. It's close but it has issues. When the loop starts
getting to the edges, it doesn't know what to do and I'm too tired to
troubleshoot it right now.
You have been warned, if you run the imageSliceNDice function that is in
driver, it will dump 49 new image files into your pygame folder. lol.
Also you will need a package called Pillow if you don't already have it.